// testeEddyData/src/app/models/cliente.model.ts
export interface Cliente {
  nome?: string
  cpf?: string
  estadocivil?: string
  pai?: string
  mae?: string
  conjuge?: string
  rg?: string
  salario?: string | number
  especie?: string
  titulodeeleitor?: string
  sexo?: string
  celular?: string
  cep?: string
  endereco?: string
  complemento?: string
  numero?: string
  bairro?: string
  email?: string
  cidade?: string
  datadenascimento?: string | Date
}